requirements.txt
fastapi==0.115.12
pydantic==2.11.3
opencv-python==4.11.0.86
pillow==11.2.1
pytesseract==0.3.13
python-multipart==0.0.20
uvicorn==0.34.2
